+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+The main changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [Unreleased]
+## [0.0.10]
 
 - Added `Element`/`AsyncElement.parent` property.
 - Fixed the `:checked` pseudo-class to also match a selected `<option>`, not just